RI-5477: Put Formik context at RDI Instance level to preserve context across pipeline management and statistics pages.
