+++ redisinsight/ui/src/components/main-router/constants/defaultRoutes.ts
@@ -15,7 +15,7 @@ import AnalyticsPage from 'uiSrc/pages/analytics'
 import TriggeredFunctionsPage from 'uiSrc/pages/triggered-functions'
 import RdiPage from 'uiSrc/pages/rdi/home'
 import RdiInstancePage from 'uiSrc/pages/rdi/instance'
-import RdiStatisticsPage from 'uiSrc/pages/rdi/statistics/StatisticsPage'
+import RdiStatisticsPage from 'uiSrc/pages/rdi/statistics'
 import PipelineManagementPage from 'uiSrc/pages/rdi/pipeline-management'
 import { ANALYTICS_ROUTES, RDI_PIPELINE_MANAGEMENT_ROUTES, TRIGGERED_FUNCTIONS_ROUTES } from './sub-routes'
 
@@ -51,7 +51,6 @@ const INSTANCE_ROUTES: IRoute[] = [
 
 const RDI_INSTANCE_ROUTES: IRoute[] = [
   {
-    pageName: PageNames.rdiStatistics,
     path: Pages.rdiStatistics(':rdiInstanceId'),
     component: RdiStatisticsPage,
   },
